+++ src/layouts/default.vue
@@ -67,7 +67,7 @@
 </template>
 
 <script>
-  import HomeHeader from '~/components/HomeHeader.vue'
+  import HomeHeader from '~/components/HomeHeader/Index.vue'
 
   export default {
     components: { HomeHeader },
