+++ src/pages/welcome/index.vue
